+++ doc/Changelog.md
@@ -1,5 +1,7 @@
 # version 0.10.0
 - derivation path struct
+- settings: anu
+- settings: proxy
 
 
 # version 0.9.1